department setup test done and fixes service methods

+++ src/modules/deparments/services/DeleteDepartmentsService.ts
@@ -1,4 +1,5 @@
 import { injectable, inject } from 'tsyringe';
+import AppError from '@shared/errors/AppError';
 import IDepartmentsRepository from '../repositories/IDepartmentsRepository';
 
 @injectable()
@@ -9,6 +10,11 @@ export default class DeleteDepartmentsService {
   ) {}
 
   public async execute(id: string): Promise<void> {
+    const department = await this.departmentsRepository.findById(id);
+
+    if (!department) {
+      throw new AppError('Department does not exists');
+    }
     await this.departmentsRepository.delete(id);
   }
 }
